backend/requirements.txt
aerich==0.7.1 ; python_version >= "3.9" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.9" and python_version < "4.0"
asyncpg==0.27.0 ; python_version >= "3.9" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.9" and python_version < "4.0"
boto3==1.26.151 ; python_version >= "3.9" and python_version < "4.0"
botocore==1.29.151 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
dictdiffer==0.9.0 ; python_version >= "3.9" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.9" and python_version < "4.0"
fastapi==0.88.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
mangum==0.17.0 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
motor==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.9" and python_version < "4.0"
phonenumbers==8.13.11 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.7 ; python_version >= "3.9" and python_version < "4.0"
pyjwt==2.7.0 ; python_version >= "3.9" and python_version < "4.0"
pymongo==4.3.3 ; python_version >= "3.9" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-decouple==3.8 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
python-http-client==3.3.7 ; python_version >= "3.9" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.9" and python_version < "4.0"
python-slugify==8.0.1 ; python_version >= "3.9" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.30.0 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
s3transfer==0.6.1 ; python_version >= "3.9" and python_version < "4.0"
sendgrid==6.10.0 ; python_version >= "3.9" and python_version < "4.0"
sentry-sdk[fastapi]==1.23.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
starkbank-ecdsa==2.2.0 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.9" and python_version < "4.0"
tomlkit==0.11.8 ; python_version >= "3.9" and python_version < "4.0"
tortoise-orm==0.19.3 ; python_version >= "3.9" and python_version < "4.0"
twilio==7.17.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.9" and python_version < "4.0"
uvicorn==0.20.0 ; python_version >= "3.9" and python_version < "4.0"
psycopg2==2.9.9 ; python_version >= "3.9" and python_version < "4.0"
fastapi-cache2[redis]; python_version >= "3.9" and python_version < "4.0"
stripe==8.7.0; python_version >= "3.9" and python_version < "4.0"

